requirements.txt
flask==3.0.3
flask-cors==4.0.1
langchain==0.2.16
langchain-openai==0.1.25
langchain-community==0.2.17
openai==1.47.0
pydantic==2.9.1
python-dotenv==1.0.1
waitress==3.0.1
redis==5.0.8
